+++ packages/nextjs/components/blockexplorer/SearchBar.tsx
@@ -13,6 +13,7 @@ export const SearchBar: React.FC<SearchBarProps> = ({ searchInput, setSearchInpu
         className="border-primary bg-base-100 text-base-content p-2 mr-2 w-full md:w-1/2 lg:w-1/3 rounded-md shadow-sm focus:outline-none focus:ring-2 focus:ring-accent"
         type="text"
         value={searchInput}
+        placeholder="Search by hash or address"
         onChange={e => setSearchInput(e.target.value)}
       />
       <button className="btn btn-primary" type="submit">
