+++ packages/nextjs/utils/scaffold-eth/block.ts
@@ -1,0 +1,18 @@
+import { BlockWithTransactions, TransactionResponse } from "@ethersproject/abstract-provider";
+
+export type TransactionWithFunction = TransactionResponse & { functionName?: string };
+
+export interface Block extends BlockWithTransactions {
+  transactions: TransactionWithFunction[];
+}
+
+interface TransactionReceipts {
+  [key: string]: {
+    contractAddress: string;
+  };
+}
+
+export interface TransactionsTableProps {
+  blocks: Block[];
+  transactionReceipts: TransactionReceipts;
+}
